Move SessionManager to session_manager.py
