+++ discovery-provider/src/conftest.py
@@ -13,7 +13,7 @@ import src.utils.db_session
 # Test fixture to mock a postgres database using an in-memory alternative
 @pytest.fixture()
 def db_mock(monkeypatch):
-    db = src.utils.db_session.SessionManager(
+    db = src.utils.session_manager.SessionManager(
         'sqlite://',
         {}
     )
